requirements.txt
opensearch-py==2.7.1
boto3==1.37.1
requests-aws4auth==1.3
pyyaml==6.0.2
requests==2.32.3
streamlit==1.41.1
aws-cdk-lib==2.186.0
constructs>=10.0.0,<11.0.0
beautifulsoup4==4.12.3
python-dotenv==1.0.0
pandas>=2.0.0
pyarrow>=14.0.0
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
atlassian-python-api
types-PyYAML
pymediainfo
requests
